Move OpenCensus instrumentation to a sub-package
