+++ go.mod
@@ -7,4 +7,5 @@ require (
 	github.com/onsi/ginkgo v1.14.1
 	github.com/onsi/gomega v1.10.2
 	go.opentelemetry.io/otel v0.13.0
+	golang.org/x/text v0.3.3 // indirect
 )
