+++ rediscensus/go.mod
@@ -1,0 +1,9 @@
+module github.com/go-redis/redisext/rediscensus
+
+go 1.15
+
+require (
+	github.com/go-redis/redis/v8 v8.3.0
+	github.com/go-redis/redisext v0.2.1
+	go.opencensus.io v0.22.5
+)
